sensores_iot/requirements.txt
kafka-python
faker
python-dotenv
psycopg2-binary